feature(ProcessPipe): Add a method for a processor to overwrite the source samplerate

+++ tests/test_graphers_render_analyzers.py
@@ -38,7 +38,8 @@ class Test_graphers_analyzers(unittest.TestCase):
 
     def tearDown(self):
         # Clean-up : delete temp file
-        os.unlink(self.temp_file.name)
+        if not PLOT:
+            os.unlink(self.temp_file.name)
 
 
 def _tests_factory(grapher_analyzers):
